Extract Block to its own module
