+++ artichoke-backend/src/extn/prelude.rs
@@ -14,6 +14,7 @@
 //!
 //! The prelude may grow over time as additional items see ubiquitous use.
 
+pub use crate::block::Block;
 pub use crate::class;
 pub use crate::class_registry::ClassRegistry;
 pub use crate::convert::RustBackedValue;
@@ -26,7 +27,7 @@ pub use crate::prelude::*;
 pub use crate::string;
 pub use crate::sys;
 pub use crate::types::{Fp, Int};
-pub use crate::value::{Block, Value};
+pub use crate::value::Value;
 
 /// Type alias for errors returned from `init` functions in
 /// [`extn`](crate::extn).
